requirements.txt
streamlit>=1.35.0
ultralytics>=8.2.48
opencv-python>=4.10.0.82
pillow>=10.3.0
numpy>=1.26.4